Position Brand section behind core values
I have to figure out a better way to solve this.

I think that maybe using the "brand" section to be relative,
and the Core Values section obsolute positioned from the
Brand section will require less CSS and a more elegant
solution to the best of my knowledge right now.