Change some of the layout
I decided to change some of the design. I felt like some of the
sections were disconnected from the overall flow of information.

I added a light bg to the product categories section to add
importance, changed the layout for the season product section, as well as the core values one. Everything else stayed the same.